Removed condition on 'ci skip'
We built a mechanism to skip building commits with a message containing 'ci skip'. The downside of the custom mechanism was that always two builds were done: a real one and a dummy one that skipped all work. Now we use the plug-in "GitHub Commit Skip SCM Behaviour", which takes care of the skips earlier in the process in Jenkins.